Increase responsiveness by caching next image
When using sxiv as a photo viewer for large photos (6000x4000), the vast
majority of time is spent in rendering (img_load -> img_open ->
imlib_image_get_data_for_reading_only -> ...).

For the normal case where the user is going to view the next image, we
can improve responsiveness by caching the next image as a kind of
readahead. We do this after we've finished rendering the current image,
so for the normal case where we navigate forwards every few seconds,
this makes navigation almost instantaneous (decreasing time from key
press to image displayed by about 60-70% in my tests).

We decide which image should next be cached based on the previous
navigation. For example, if you are at index 400 and you navigate to
410, we try to cache 420. If you are at index 355 and you navigate to
354, we try to cache 353. This increases the cache hit rate regardless
of the navigation which the user is performing. Even in the case where
the cache is missed, usually responsiveness is no worse than usual,
since it usually takes less time for the image to be cached than it
takes for the user to perform a subsequent navigation.